test: cover fight-state helper modules
