+++ src/features/fight-state/fightReducer.test.ts
@@ -1,15 +1,194 @@
 import { describe, expect, it } from 'vitest';
 
 import { bossMap, bossSequenceMap, charmMap, DEFAULT_BOSS_ID } from '../../data';
+import { applyLogUpdate, resolveFightCompletion } from './reducer/actions';
+import {
+  appendEventAggregates,
+  createEmptyAggregates,
+  deriveDamageLogAggregates,
+} from './reducer/aggregates';
 import {
   MAX_NOTCH_LIMIT,
   MAX_OVERCHARM_OVERFLOW,
   MIN_NOTCH_LIMIT,
+  clampCharmSelection,
+  clampNotchLimit,
   createInitialState,
-  fightReducer,
-} from './fightReducer';
+} from './reducer/build';
+import { ensureSequenceState, toSequenceStageKey } from './reducer/sequence';
+import type { AttackEvent, FightState } from './reducer/types';
+import { fightReducer } from './fightReducer';
 
 const getCharmCost = (id: string) => charmMap.get(id)?.cost ?? 0;
+
+describe('fight reducer helper modules', () => {
+  describe('build helpers', () => {
+    it('clamps charm selections to the configured notch limit', () => {
+      const richSelection = ['shaman-stone', 'spell-twister', 'quick-slash'];
+      const clamped = clampCharmSelection(richSelection, 5);
+
+      const totalCost = clamped.reduce((total, id) => total + getCharmCost(id), 0);
+
+      expect(totalCost).toBeLessThanOrEqual(5 + MAX_OVERCHARM_OVERFLOW);
+      expect(clamped).toEqual(richSelection);
+    });
+
+    it('respects the minimum and maximum notch limits', () => {
+      expect(clampNotchLimit(1)).toBe(MIN_NOTCH_LIMIT);
+      expect(clampNotchLimit(30)).toBe(MAX_NOTCH_LIMIT);
+      expect(clampNotchLimit(7)).toBe(7);
+    });
+  });
+
+  describe('aggregates helpers', () => {
+    it('derives aggregates from logged events', () => {
+      const events = [
+        { id: 'a', label: 'A', damage: 5, category: 'nail', timestamp: 100 },
+        { id: 'b', label: 'B', damage: 7, category: 'spell', timestamp: 200 },
+        { id: 'c', label: 'C', damage: 3, category: 'nail-art', timestamp: 150 },
+      ] satisfies AttackEvent[];
+
+      const aggregates = deriveDamageLogAggregates(events);
+
+      expect(aggregates.totalDamage).toBe(15);
+      expect(aggregates.attacksLogged).toBe(3);
+      expect(aggregates.firstAttackTimestamp).toBe(100);
+      expect(aggregates.lastAttackTimestamp).toBe(200);
+    });
+  });
+
+  describe('actions helpers', () => {
+    it('resolves fight completion when damage reaches the target HP', () => {
+      const state = createInitialState();
+      const targetHp = bossMap.get(state.selectedBossId)?.hp ?? 0;
+      const event: AttackEvent = {
+        id: 'finisher',
+        label: 'Finisher',
+        damage: targetHp,
+        category: 'nail',
+        timestamp: 1_000,
+      };
+      const log = [event];
+      const aggregates = appendEventAggregates(createEmptyAggregates(), event);
+
+      const completion = resolveFightCompletion(state, log, aggregates);
+
+      expect(completion.fightEndTimestamp).toBe(event.timestamp);
+      expect(completion.fightManuallyEnded).toBe(false);
+    });
+
+    it('preserves a pre-existing completion timestamp when requested', () => {
+      const state: FightState = {
+        ...createInitialState(),
+        fightEndTimestamp: 2_000,
+        fightManuallyEnded: false,
+      };
+      const event: AttackEvent = {
+        id: 'follow-up',
+        label: 'Follow Up',
+        damage: 1,
+        category: 'nail',
+        timestamp: 2_500,
+      };
+      const log = [event];
+      const aggregates = appendEventAggregates(createEmptyAggregates(), event);
+
+      const completion = resolveFightCompletion(state, log, aggregates, {
+        preserveEndTimestamp: true,
+      });
+
+      expect(completion.fightEndTimestamp).toBe(2_000);
+      expect(completion.fightManuallyEnded).toBe(false);
+    });
+
+    it('applies log updates across active sequence state', () => {
+      const event: AttackEvent = {
+        id: 'stage-event',
+        label: 'Stage Event',
+        damage: 10,
+        category: 'nail',
+        timestamp: 3_000,
+      };
+      const aggregates = appendEventAggregates(createEmptyAggregates(), event);
+      const sequenceId = 'test-sequence';
+      const state: FightState = {
+        ...createInitialState(),
+        activeSequenceId: sequenceId,
+        sequenceIndex: 1,
+      };
+      const fightCompletion = {
+        fightEndTimestamp: 4_000,
+        fightManuallyEnded: false,
+      };
+      const fightStart = { timestamp: 2_000, manuallyStarted: true };
+
+      const updated = applyLogUpdate(
+        state,
+        [event],
+        aggregates,
+        [event],
+        fightCompletion,
+        fightStart,
+      );
+
+      const key = toSequenceStageKey(sequenceId, 1);
+
+      expect(updated.damageLog).toEqual([event]);
+      expect(updated.damageLogAggregates).toEqual(aggregates);
+      expect(updated.damageLogVersion).toBe(state.damageLogVersion + 1);
+      expect(updated.redoStack).toEqual([event]);
+      expect(updated.fightEndTimestamp).toBe(4_000);
+      expect(updated.fightManuallyEnded).toBe(false);
+      expect(updated.fightStartTimestamp).toBe(2_000);
+      expect(updated.fightManuallyStarted).toBe(true);
+      expect(updated.sequenceLogs[key]).toEqual([event]);
+      expect(updated.sequenceLogAggregates[key]).toEqual(aggregates);
+      expect(updated.sequenceRedoStacks[key]).toEqual([event]);
+      expect(updated.sequenceFightStartTimestamps[key]).toBe(2_000);
+      expect(updated.sequenceManualStartFlags[key]).toBe(true);
+      expect(updated.sequenceFightEndTimestamps[key]).toBe(4_000);
+      expect(updated.sequenceManualEndFlags[key]).toBe(false);
+    });
+  });
+});
+
+describe('sequence helpers', () => {
+  it('clamps invalid sequence indices and reloads stored logs', () => {
+    const masterSequence = bossSequenceMap.get('pantheon-of-the-master');
+    if (!masterSequence) {
+      throw new Error('Expected pantheon-of-the-master sequence to be defined for tests');
+    }
+
+    let state = createInitialState();
+    state = fightReducer(state, {
+      type: 'startSequence',
+      sequenceId: masterSequence.id,
+    });
+
+    const stageDamage = Math.max(1, Math.floor(masterSequence.entries[0].target.hp / 2));
+    state = fightReducer(state, {
+      type: 'logAttack',
+      id: 'stage-0',
+      label: 'Stage 0 Hit',
+      damage: stageDamage,
+      category: 'nail',
+      timestamp: 1_000,
+    });
+
+    const mutated: FightState = {
+      ...state,
+      damageLog: [],
+      damageLogAggregates: deriveDamageLogAggregates([]),
+      sequenceIndex: -5,
+    };
+
+    const ensured = ensureSequenceState(mutated);
+
+    expect(ensured.sequenceIndex).toBe(0);
+    expect(ensured.damageLog).toHaveLength(1);
+    expect(ensured.damageLogAggregates.totalDamage).toBe(stageDamage);
+  });
+});
 
 describe('fightReducer notch limits', () => {
   it('clamps charm selections to the configured notch limit', () => {
