+++ src/features/fight-state/fightReducer.ts
@@ -1,612 +1,52 @@
+import type { AttackEvent, FightAction, FightState } from './reducer/types';
+import { CUSTOM_BOSS_ID } from './reducer/types';
 import {
-  DEFAULT_BOSS_ID,
-  DEFAULT_CUSTOM_HP,
-  bossMap,
-  bossSequenceMap,
-  charmMap,
-  resolveSequenceEntries,
-  nailUpgrades,
-  spells,
-} from '../../data';
-import type { BossSequenceEntry } from '../../data';
+  appendEventAggregates,
+  createEmptyAggregates,
+  removeLastEventAggregates,
+} from './reducer/aggregates';
+import { applyLogUpdate, resolveFightCompletion } from './reducer/actions';
+import { clampCharmSelection, clampNotchLimit } from './reducer/build';
+import {
+  ensureSequenceState,
+  exitSequence,
+  filterSequenceRecords,
+  getResolvedSequenceEntries,
+  loadSequenceStage,
+  persistCurrentSequenceStage,
+  toSequenceStageKey,
+  withDamageLog,
+} from './reducer/sequence';
 
-export type AttackCategory = 'nail' | 'spell' | 'nail-art' | 'charm';
-export type SpellLevel = 'none' | 'base' | 'upgrade';
-
-export interface AttackEvent {
-  id: string;
-  label: string;
-  damage: number;
-  category: AttackCategory;
-  timestamp: number;
-  soulCost?: number;
-}
-
-export interface DamageLogAggregates {
-  totalDamage: number;
-  attacksLogged: number;
-  firstAttackTimestamp: number | null;
-  lastAttackTimestamp: number | null;
-}
-
-const createEmptyAggregates = (): DamageLogAggregates => ({
-  totalDamage: 0,
-  attacksLogged: 0,
-  firstAttackTimestamp: null,
-  lastAttackTimestamp: null,
-});
-
-const cloneAggregates = (aggregates: DamageLogAggregates): DamageLogAggregates => ({
-  totalDamage: aggregates.totalDamage,
-  attacksLogged: aggregates.attacksLogged,
-  firstAttackTimestamp: aggregates.firstAttackTimestamp,
-  lastAttackTimestamp: aggregates.lastAttackTimestamp,
-});
-
-export const deriveDamageLogAggregates = (
-  damageLog: AttackEvent[],
-): DamageLogAggregates => {
-  if (damageLog.length === 0) {
-    return createEmptyAggregates();
-  }
-
-  let totalDamage = 0;
-  let firstAttackTimestamp: number | null = null;
-  let lastAttackTimestamp: number | null = null;
-
-  for (const event of damageLog) {
-    totalDamage += event.damage;
-
-    if (firstAttackTimestamp === null || event.timestamp < firstAttackTimestamp) {
-      firstAttackTimestamp = event.timestamp;
-    }
-
-    if (lastAttackTimestamp === null || event.timestamp > lastAttackTimestamp) {
-      lastAttackTimestamp = event.timestamp;
-    }
-  }
-
-  return {
-    totalDamage,
-    attacksLogged: damageLog.length,
-    firstAttackTimestamp,
-    lastAttackTimestamp,
-  };
-};
-
-const appendEventAggregates = (
-  aggregates: DamageLogAggregates,
-  event: AttackEvent,
-): DamageLogAggregates => {
-  const isFirstEvent = aggregates.attacksLogged === 0;
-  const firstAttackTimestamp = isFirstEvent
-    ? event.timestamp
-    : Math.min(aggregates.firstAttackTimestamp ?? event.timestamp, event.timestamp);
-  const lastAttackTimestamp = isFirstEvent
-    ? event.timestamp
-    : Math.max(aggregates.lastAttackTimestamp ?? event.timestamp, event.timestamp);
-
-  return {
-    totalDamage: aggregates.totalDamage + event.damage,
-    attacksLogged: aggregates.attacksLogged + 1,
-    firstAttackTimestamp,
-    lastAttackTimestamp,
-  };
-};
-
-const removeLastEventAggregates = (
-  nextDamageLog: AttackEvent[],
-  removedEvent: AttackEvent,
-  aggregates: DamageLogAggregates,
-): DamageLogAggregates => {
-  const nextAttacksLogged = Math.max(0, aggregates.attacksLogged - 1);
-  if (nextAttacksLogged === 0) {
-    return createEmptyAggregates();
-  }
-
-  const firstAttackTimestamp = nextDamageLog[0]?.timestamp ?? null;
-  const lastAttackTimestamp = nextDamageLog[nextDamageLog.length - 1]?.timestamp ?? null;
-
-  return {
-    totalDamage: aggregates.totalDamage - removedEvent.damage,
-    attacksLogged: nextAttacksLogged,
-    firstAttackTimestamp,
-    lastAttackTimestamp,
-  };
-};
-
-export interface BuildState {
-  nailUpgradeId: string;
-  activeCharmIds: string[];
-  spellLevels: Record<string, SpellLevel>;
-  notchLimit: number;
-}
-
-export interface FightState {
-  selectedBossId: string;
-  customTargetHp: number;
-  build: BuildState;
-  damageLog: AttackEvent[];
-  damageLogAggregates: DamageLogAggregates;
-  damageLogVersion: number;
-  redoStack: AttackEvent[];
-  activeSequenceId: string | null;
-  sequenceIndex: number;
-  sequenceLogs: Partial<Record<string, AttackEvent[]>>;
-  sequenceLogAggregates: Partial<Record<string, DamageLogAggregates>>;
-  sequenceRedoStacks: Partial<Record<string, AttackEvent[]>>;
-  sequenceConditions: Partial<Record<string, Record<string, boolean>>>;
-  fightStartTimestamp: number | null;
-  fightManuallyStarted: boolean;
-  fightEndTimestamp: number | null;
-  fightManuallyEnded: boolean;
-  sequenceFightStartTimestamps: Partial<Record<string, number | null>>;
-  sequenceManualStartFlags: Partial<Record<string, boolean>>;
-  sequenceFightEndTimestamps: Partial<Record<string, number | null>>;
-  sequenceManualEndFlags: Partial<Record<string, boolean>>;
-}
-
-export interface AttackInput {
-  id: string;
-  label: string;
-  damage: number;
-  category: AttackCategory;
-  soulCost?: number;
-  timestamp?: number;
-}
-
-export const CUSTOM_BOSS_ID = 'custom';
-
-export type FightAction =
-  | { type: 'selectBoss'; bossId: string }
-  | { type: 'setCustomTargetHp'; hp: number }
-  | { type: 'setNailUpgrade'; nailUpgradeId: string }
-  | { type: 'setActiveCharms'; charmIds: string[] }
-  | {
-      type: 'updateActiveCharms';
-      updater: (charmIds: string[]) => string[];
-    }
-  | { type: 'setCharmNotchLimit'; notchLimit: number }
-  | { type: 'setSpellLevel'; spellId: string; level: SpellLevel }
-  | {
-      type: 'logAttack';
-      timestamp: number;
-      label: string;
-      id: string;
-      damage: number;
-      category: AttackCategory;
-      soulCost?: number;
-    }
-  | { type: 'undoLastAttack' }
-  | { type: 'redoLastAttack' }
-  | { type: 'resetLog' }
-  | { type: 'resetSequence' }
-  | { type: 'startFight'; timestamp: number }
-  | { type: 'endFight'; timestamp: number }
-  | { type: 'startSequence'; sequenceId: string }
-  | { type: 'stopSequence' }
-  | { type: 'setSequenceStage'; index: number }
-  | { type: 'advanceSequence' }
-  | { type: 'rewindSequence' }
-  | {
-      type: 'setSequenceCondition';
-      sequenceId: string;
-      conditionId: string;
-      enabled: boolean;
-    };
-
-export const toSequenceStageKey = (sequenceId: string, index: number) =>
-  `${sequenceId}#${index}`;
-
-export const initialSpellLevels = (): Record<string, SpellLevel> => {
-  const levels: Record<string, SpellLevel> = {};
-  for (const spell of spells) {
-    levels[spell.id] = 'base';
-  }
-  return levels;
-};
-
-export const MIN_NOTCH_LIMIT = 3;
-export const MAX_NOTCH_LIMIT = 11;
-export const MAX_OVERCHARM_OVERFLOW = 5;
-
-const clampNotchLimit = (value: number) =>
-  Math.min(MAX_NOTCH_LIMIT, Math.max(MIN_NOTCH_LIMIT, Math.round(value)));
-
-const getCharmCost = (charmId: string) => charmMap.get(charmId)?.cost ?? 0;
-
-const clampCharmSelection = (charmIds: string[], notchLimit: number) => {
-  if (notchLimit <= 0) {
-    return [];
-  }
-
-  const uniqueOrdered = charmIds.filter((id, index) => charmIds.indexOf(id) === index);
-  const selected: string[] = [];
-  let totalCost = 0;
-  let hasOverflow = false;
-
-  for (const id of uniqueOrdered) {
-    const cost = getCharmCost(id);
-    const nextCost = totalCost + cost;
-    if (nextCost > notchLimit) {
-      if (hasOverflow || nextCost > notchLimit + MAX_OVERCHARM_OVERFLOW) {
-        continue;
-      }
-      hasOverflow = true;
-    }
-
-    selected.push(id);
-    totalCost = nextCost;
-  }
-
-  return selected;
-};
-
-export const createInitialState = (): FightState => ({
-  selectedBossId: DEFAULT_BOSS_ID,
-  customTargetHp: DEFAULT_CUSTOM_HP,
-  build: {
-    nailUpgradeId: nailUpgrades[0]?.id ?? 'old-nail',
-    activeCharmIds: [],
-    spellLevels: initialSpellLevels(),
-    notchLimit: MAX_NOTCH_LIMIT,
-  },
-  damageLog: [],
-  damageLogAggregates: createEmptyAggregates(),
-  damageLogVersion: 0,
-  redoStack: [],
-  activeSequenceId: null,
-  sequenceIndex: 0,
-  sequenceLogs: {},
-  sequenceLogAggregates: {},
-  sequenceRedoStacks: {},
-  sequenceConditions: {},
-  fightStartTimestamp: null,
-  fightManuallyStarted: false,
-  fightEndTimestamp: null,
-  fightManuallyEnded: false,
-  sequenceFightStartTimestamps: {},
-  sequenceManualStartFlags: {},
-  sequenceFightEndTimestamps: {},
-  sequenceManualEndFlags: {},
-});
+export type {
+  AttackCategory,
+  AttackEvent,
+  AttackInput,
+  BuildState,
+  DamageLogAggregates,
+  FightAction,
+  FightState,
+  SpellLevel,
+} from './reducer/types';
+export { CUSTOM_BOSS_ID } from './reducer/types';
+export {
+  createInitialState,
+  ensureSpellLevels,
+  initialSpellLevels,
+  MAX_NOTCH_LIMIT,
+  MAX_OVERCHARM_OVERFLOW,
+  MIN_NOTCH_LIMIT,
+} from './reducer/build';
+export { deriveDamageLogAggregates } from './reducer/aggregates';
+export {
+  ensureSequenceState,
+  exitSequence,
+  loadSequenceStage,
+  persistCurrentSequenceStage,
+  toSequenceStageKey,
+} from './reducer/sequence';
 
 export const isCustomBoss = (bossId: string) => bossId === CUSTOM_BOSS_ID;
-
-const getTargetHp = (state: FightState) =>
-  isCustomBoss(state.selectedBossId)
-    ? Math.max(1, Math.round(state.customTargetHp))
-    : (bossMap.get(state.selectedBossId)?.hp ?? DEFAULT_CUSTOM_HP);
-
-const withDamageLog = (
-  state: FightState,
-  damageLog: AttackEvent[],
-  aggregates: DamageLogAggregates,
-): FightState => ({
-  ...state,
-  damageLog,
-  damageLogAggregates: cloneAggregates(aggregates),
-  damageLogVersion: state.damageLogVersion + 1,
-});
-
-const resolveFightCompletion = (
-  state: FightState,
-  damageLog: AttackEvent[],
-  aggregates: DamageLogAggregates,
-  options?: {
-    preserveEndTimestamp?: boolean;
-  },
-) => {
-  const preserveEndTimestamp = options?.preserveEndTimestamp ?? false;
-  const totalDamage = aggregates.totalDamage;
-  const targetHp = getTargetHp(state);
-
-  if (preserveEndTimestamp && state.fightEndTimestamp !== null) {
-    return {
-      fightEndTimestamp: state.fightEndTimestamp,
-      fightManuallyEnded: state.fightManuallyEnded,
-    };
-  }
-
-  if (state.fightManuallyEnded) {
-    return {
-      fightEndTimestamp: null,
-      fightManuallyEnded: false,
-    };
-  }
-
-  if (damageLog.length === 0) {
-    return {
-      fightEndTimestamp: null,
-      fightManuallyEnded: false,
-    };
-  }
-
-  if (totalDamage >= targetHp) {
-    const lastEvent = damageLog[damageLog.length - 1];
-    return {
-      fightEndTimestamp: lastEvent.timestamp,
-      fightManuallyEnded: false,
-    };
-  }
-
-  return {
-    fightEndTimestamp: null,
-    fightManuallyEnded: false,
-  };
-};
-
-export const persistCurrentSequenceStage = (state: FightState): FightState => {
-  if (!state.activeSequenceId) {
-    return state;
-  }
-
-  const key = toSequenceStageKey(state.activeSequenceId, state.sequenceIndex);
-  return {
-    ...state,
-    sequenceLogs: {
-      ...state.sequenceLogs,
-      [key]: [...state.damageLog],
-    },
-    sequenceLogAggregates: {
-      ...state.sequenceLogAggregates,
-      [key]: cloneAggregates(state.damageLogAggregates),
-    },
-    sequenceRedoStacks: {
-      ...state.sequenceRedoStacks,
-      [key]: [...state.redoStack],
-    },
-    sequenceFightStartTimestamps: {
-      ...state.sequenceFightStartTimestamps,
-      [key]: state.fightStartTimestamp,
-    },
-    sequenceManualStartFlags: {
-      ...state.sequenceManualStartFlags,
-      [key]: state.fightManuallyStarted,
-    },
-    sequenceFightEndTimestamps: {
-      ...state.sequenceFightEndTimestamps,
-      [key]: state.fightEndTimestamp,
-    },
-    sequenceManualEndFlags: {
-      ...state.sequenceManualEndFlags,
-      [key]: state.fightManuallyEnded,
-    },
-  };
-};
-
-const filterSequenceRecords = <T>(
-  records: Partial<Record<string, T>>,
-  sequenceId: string,
-) => {
-  const prefix = `${sequenceId}#`;
-  return Object.fromEntries(
-    Object.entries(records).filter(([key]) => !key.startsWith(prefix)),
-  ) as Partial<Record<string, T>>;
-};
-
-const getResolvedSequenceEntries = (
-  state: FightState,
-  sequenceId: string,
-): { sequence: ReturnType<typeof bossSequenceMap.get>; entries: BossSequenceEntry[] } => {
-  const sequence = bossSequenceMap.get(sequenceId);
-  if (!sequence) {
-    return { sequence: undefined, entries: [] };
-  }
-  const overrides = state.sequenceConditions[sequenceId];
-  const entries = resolveSequenceEntries(sequence, overrides);
-  return { sequence, entries };
-};
-
-export const loadSequenceStage = (
-  state: FightState,
-  sequenceId: string,
-  index: number,
-): FightState => {
-  const { sequence, entries } = getResolvedSequenceEntries(state, sequenceId);
-  if (!sequence || entries.length === 0) {
-    return {
-      ...state,
-      activeSequenceId: null,
-      sequenceIndex: 0,
-    };
-  }
-
-  const clampedIndex = Math.max(0, Math.min(index, entries.length - 1));
-  const key = toSequenceStageKey(sequenceId, clampedIndex);
-  const storedLog = state.sequenceLogs[key] ?? [];
-  const storedRedo = state.sequenceRedoStacks[key] ?? [];
-  const storedStartTimestamp = state.sequenceFightStartTimestamps[key] ?? null;
-  const storedManualStart = state.sequenceManualStartFlags[key] ?? false;
-  const storedEndTimestamp = state.sequenceFightEndTimestamps[key] ?? null;
-  const storedManualEnd = state.sequenceManualEndFlags[key] ?? false;
-  const storedAggregates = state.sequenceLogAggregates[key];
-  const nextTarget = entries[clampedIndex].target;
-
-  const aggregates = storedAggregates ?? deriveDamageLogAggregates(storedLog);
-  const baseState = withDamageLog(state, [...storedLog], aggregates);
-  const nextSequenceLogAggregates =
-    storedAggregates === undefined
-      ? {
-          ...state.sequenceLogAggregates,
-          [key]: cloneAggregates(aggregates),
-        }
-      : state.sequenceLogAggregates;
-
-  return {
-    ...baseState,
-    activeSequenceId: sequenceId,
-    sequenceIndex: clampedIndex,
-    selectedBossId: nextTarget.id,
-    redoStack: [...storedRedo],
-    fightStartTimestamp: storedStartTimestamp,
-    fightManuallyStarted: storedManualStart,
-    fightEndTimestamp: storedEndTimestamp,
-    fightManuallyEnded: storedManualEnd,
-    sequenceLogAggregates: nextSequenceLogAggregates,
-  };
-};
-
-export const exitSequence = (state: FightState): FightState => {
-  if (!state.activeSequenceId) {
-    return state;
-  }
-
-  const persisted = persistCurrentSequenceStage(state);
-  return {
-    ...persisted,
-    activeSequenceId: null,
-    sequenceIndex: 0,
-  };
-};
-
-export const applyLogUpdate = (
-  state: FightState,
-  damageLog: AttackEvent[],
-  aggregates: DamageLogAggregates,
-  redoStack: AttackEvent[],
-  fightCompletion: { fightEndTimestamp: number | null; fightManuallyEnded: boolean },
-  fightStart: { timestamp: number | null; manuallyStarted: boolean },
-): FightState => {
-  const baseState = withDamageLog(state, damageLog, aggregates);
-
-  if (!state.activeSequenceId) {
-    return {
-      ...baseState,
-      redoStack,
-      fightEndTimestamp: fightCompletion.fightEndTimestamp,
-      fightManuallyEnded: fightCompletion.fightManuallyEnded,
-      fightStartTimestamp: fightStart.timestamp,
-      fightManuallyStarted: fightStart.manuallyStarted,
-    };
-  }
-
-  const key = toSequenceStageKey(state.activeSequenceId, state.sequenceIndex);
-  return {
-    ...baseState,
-    redoStack,
-    fightEndTimestamp: fightCompletion.fightEndTimestamp,
-    fightManuallyEnded: fightCompletion.fightManuallyEnded,
-    fightStartTimestamp: fightStart.timestamp,
-    fightManuallyStarted: fightStart.manuallyStarted,
-    sequenceLogs: {
-      ...state.sequenceLogs,
-      [key]: damageLog,
-    },
-    sequenceLogAggregates: {
-      ...state.sequenceLogAggregates,
-      [key]: cloneAggregates(aggregates),
-    },
-    sequenceRedoStacks: {
-      ...state.sequenceRedoStacks,
-      [key]: redoStack,
-    },
-    sequenceFightStartTimestamps: {
-      ...state.sequenceFightStartTimestamps,
-      [key]: fightStart.timestamp,
-    },
-    sequenceManualStartFlags: {
-      ...state.sequenceManualStartFlags,
-      [key]: fightStart.manuallyStarted,
-    },
-    sequenceFightEndTimestamps: {
-      ...state.sequenceFightEndTimestamps,
-      [key]: fightCompletion.fightEndTimestamp,
-    },
-    sequenceManualEndFlags: {
-      ...state.sequenceManualEndFlags,
-      [key]: fightCompletion.fightManuallyEnded,
-    },
-  };
-};
-
-const ensureNotchLimit = (state: FightState): FightState => {
-  const desiredLimit =
-    typeof state.build.notchLimit === 'number'
-      ? clampNotchLimit(state.build.notchLimit)
-      : MAX_NOTCH_LIMIT;
-
-  const clampedCharms = clampCharmSelection(state.build.activeCharmIds, desiredLimit);
-  const hasSameLimit = state.build.notchLimit === desiredLimit;
-  const hasSameCharms =
-    clampedCharms.length === state.build.activeCharmIds.length &&
-    clampedCharms.every((id, index) => state.build.activeCharmIds[index] === id);
-
-  if (hasSameLimit && hasSameCharms) {
-    return state;
-  }
-
-  return {
-    ...state,
-    build: {
-      ...state.build,
-      notchLimit: desiredLimit,
-      activeCharmIds: clampedCharms,
-    },
-  };
-};
-
-export const ensureSpellLevels = (state: FightState): FightState => {
-  if (Object.keys(state.build.spellLevels).length === 0) {
-    return ensureNotchLimit({
-      ...state,
-      build: {
-        ...state.build,
-        spellLevels: initialSpellLevels(),
-      },
-    });
-  }
-
-  return ensureNotchLimit(state);
-};
-
-export const ensureSequenceState = (state: FightState): FightState => {
-  if (!state.activeSequenceId) {
-    return state;
-  }
-
-  const { sequence, entries } = getResolvedSequenceEntries(state, state.activeSequenceId);
-  if (!sequence || entries.length === 0) {
-    return {
-      ...state,
-      activeSequenceId: null,
-      sequenceIndex: 0,
-    };
-  }
-
-  const clampedIndex = Math.max(0, Math.min(state.sequenceIndex, entries.length - 1));
-  const key = toSequenceStageKey(state.activeSequenceId, clampedIndex);
-  const damageLog = state.sequenceLogs[key] ?? [];
-  const redoStack = state.sequenceRedoStacks[key] ?? [];
-  const fightStartTimestamp = state.sequenceFightStartTimestamps[key] ?? null;
-  const fightManuallyStarted = state.sequenceManualStartFlags[key] ?? false;
-  const fightEndTimestamp = state.sequenceFightEndTimestamps[key] ?? null;
-  const fightManuallyEnded = state.sequenceManualEndFlags[key] ?? false;
-  const targetId = entries[clampedIndex].target.id;
-  const storedAggregates = state.sequenceLogAggregates[key];
-  const aggregates = storedAggregates ?? deriveDamageLogAggregates(damageLog);
-
-  const baseState = withDamageLog(state, [...damageLog], aggregates);
-  const nextSequenceLogAggregates =
-    storedAggregates === undefined
-      ? {
-          ...state.sequenceLogAggregates,
-          [key]: cloneAggregates(aggregates),
-        }
-      : state.sequenceLogAggregates;
-
-  return {
-    ...baseState,
-    sequenceIndex: clampedIndex,
-    selectedBossId: targetId,
-    redoStack: [...redoStack],
-    fightStartTimestamp,
-    fightManuallyStarted,
-    fightEndTimestamp,
-    fightManuallyEnded,
-    sequenceLogAggregates: nextSequenceLogAggregates,
-  };
-};
 
 export const fightReducer = (state: FightState, action: FightAction): FightState => {
   switch (action.type) {
@@ -936,7 +376,7 @@ export const fightReducer = (state: FightState, action: FightAction): FightState
     case 'stopSequence':
       return exitSequence(state);
     case 'setSequenceCondition': {
-      const sequence = bossSequenceMap.get(action.sequenceId);
+      const { sequence } = getResolvedSequenceEntries(state, action.sequenceId);
       if (!sequence) {
         return state;
       }
